meteo_project/requirements.txt
Django>=4.2
djangorestframework>=3.14
requests>=2.31.0
